Split out class
