+++ phpcs.xml
@@ -1,0 +1,9 @@
+<?xml version="1.0"?>
+<ruleset name="PluginStandard">
+    <description>Plugin Coding Standard</description>
+    <rule ref="Yoast">
+        <exclude name="Yoast.NamingConventions.NamespaceName"/>
+    </rule>
+    <exclude-pattern>vendor</exclude-pattern>
+    <exclude-pattern>tests</exclude-pattern>
+</ruleset>
